消除告警

+++ source/co_o1_scheduler.cpp
@@ -17,13 +17,13 @@ void co_o1_scheduler::add_ctx(co_ctx* ctx)
     {
         min_priority__ = ctx->priority();
     }
-    CO_DEBUG("add ctx %s %p , state: %d\n", ctx->config().name.c_str(), ctx, ctx->state());
+    CO_DEBUG("add ctx %s %p , state: %d\n", ctx->config().name.c_str(), ctx, (int)ctx->state());
 }
 
 void co_o1_scheduler::remove_ctx(co_ctx* ctx)
 {
     std::lock_guard<std::mutex> lock(mu_all_ctx__);
-    CO_DEBUG("remove ctx %s %p , state: %d", ctx->config().name.c_str(), ctx, ctx->state());
+    CO_DEBUG("remove ctx %s %p , state: %d", ctx->config().name.c_str(), ctx, (int)ctx->state());
     // 此处不能断言 curr__ != ctx，因为在最后清理所有的ctx的时候，可以删除当前ctx
     all_ctx__[ctx->priority()].remove(ctx);
 }
@@ -31,7 +31,7 @@ void co_o1_scheduler::remove_ctx(co_ctx* ctx)
 co_ctx* co_o1_scheduler::choose_ctx()
 {
     std::lock_guard<std::mutex> lock(mu_all_ctx__);
-    for (int i = min_priority__; i < all_ctx__.size(); ++i)
+    for (unsigned int i = min_priority__; i < all_ctx__.size(); ++i)
     {
         if (all_ctx__[i].empty())
         {
@@ -71,7 +71,7 @@ size_t co_o1_scheduler::count() const
     std::lock_guard<std::mutex> lock(mu_all_ctx__);
 
     size_t ret = 0;
-    for (int i = 0; i < all_ctx__.size(); ++i)
+    for (unsigned int i = 0; i < all_ctx__.size(); ++i)
     {
         ret += all_ctx__[i].size();
     }
@@ -88,7 +88,7 @@ co_ctx* co_o1_scheduler::current_ctx() const
 bool co_o1_scheduler::can_schedule() const
 {
     std::lock_guard<std::mutex> lock(mu_all_ctx__);
-    for (int i = min_priority__; i < all_ctx__.size(); ++i)
+    for (unsigned int i = min_priority__; i < all_ctx__.size(); ++i)
     {
         for (auto& ctx : all_ctx__[i])
         {
